+++ modules/TargetInfo.lua
@@ -1592,7 +1592,7 @@ function IceTargetInfo.prototype:UpdateRaidTargetIcon()
 
 	local index = GetRaidTargetIndex(self.unit);
 
-	if (index and (index > 0)) then
+	if index then
 		SetRaidTargetIconTexture(self.frame.raidIcon.icon, index)
 		self.frame.raidIcon:Show()
 	else