+++ changelog.md
@@ -6,6 +6,7 @@
 - Fix target buffs and debuffs not showing durations or stack counts in combat with some enemies in 12.0.
 - Fix Secret error in 12.0 while in Zereth Mortis around players or enemies who are in combat.
 - Re-enable mouseover buff/debuff info on Info modules. This was disabled previously due to it triggering Secret errors internal to Blizzard's code. It seems to not be doing that now.
+- Fix errors in TargetInfo, TargetHealth, and FocusHealth when targeting a unit with a raid icon set.
 
 ## v1.16.4